requirements.txt
geopy==2.4.1
joblib==1.5.0
matplotlib==3.10.3
numpy==2.3.1
ollama==0.5.1
pandas==2.3.0
pdf2image==1.17.0
pdfplumber==0.11.6
psycopg2_binary==2.9.10
pycountry==24.6.1
pytesseract==0.3.13
python-dotenv==1.1.1
rapidfuzz==3.13.0
scikit_learn==1.7.0
sentence_transformers==4.1.0
shap==0.48.0
xgboost==3.0.2
